Reorganizacija user servisa
