+++ internal/user/auth_repo.go
@@ -1,0 +1,181 @@
+package user
+
+import (
+	"bytes"
+	"crypto/sha256"
+	"database/sql"
+	"fmt"
+	"time"
+
+	"github.com/golang-jwt/jwt/v5"
+)
+
+type User struct {
+	email          string
+	hashedPassword []byte
+	salt           []byte
+}
+
+func (s *Server) GetUserByEmail(email string) (*User, error) {
+	query := `
+        SELECT email, password, salt_password FROM employees WHERE email = $1
+        UNION ALL
+        SELECT email, password, salt_password FROM clients WHERE email = $1
+        LIMIT 1
+    `
+	var user User
+	err := s.database.QueryRow(query, email).Scan(&user.email, &user.hashedPassword, &user.salt)
+	if err == sql.ErrNoRows {
+		return nil, nil
+	}
+	if err != nil {
+		return nil, err
+	}
+	return &user, nil
+}
+
+func (s *Server) rotateRefreshToken(tx *sql.Tx, email string, oldHash, newHash []byte, newExpiry time.Time) error {
+	var storedHash []byte
+	err := tx.QueryRow(`
+        SELECT hashed_token FROM refresh_tokens
+        WHERE email = $1 AND revoked = FALSE AND valid_until > now()
+        FOR UPDATE
+    `, email).Scan(&storedHash)
+	if err != nil {
+		return fmt.Errorf("refresh token not found or expired: %w", err)
+	}
+
+	if !bytes.Equal(storedHash, oldHash) {
+		_, err := tx.Exec(`UPDATE refresh_tokens SET revoked = TRUE WHERE email = $1`, email)
+		if err != nil {
+			_ = tx.Rollback()
+			return fmt.Errorf("failed to revoke tokens: %w", err)
+		}
+		_ = tx.Commit()
+		return fmt.Errorf("token mismatch: possible reuse attack")
+	}
+
+	_, err = tx.Exec(`
+        UPDATE refresh_tokens
+        SET hashed_token = $1, valid_until = $2, revoked = FALSE
+        WHERE email = $3
+    `, newHash, newExpiry, email)
+	return err
+}
+
+func (s *Server) InsertRefreshToken(token string) error {
+	parsed, _, err := jwt.NewParser().ParseUnverified(token, &jwt.RegisteredClaims{})
+	if err != nil {
+		return fmt.Errorf("parsing token: %w", err)
+	}
+
+	email, err := parsed.Claims.GetSubject()
+	if err != nil {
+		return fmt.Errorf("getting subject: %w", err)
+	}
+
+	expiry, err := parsed.Claims.GetExpirationTime()
+	if err != nil {
+		return fmt.Errorf("getting expiry: %w", err)
+	}
+	hasher := sha256.New()
+	hasher.Write([]byte(token))
+	hashed_token := hasher.Sum(nil)
+	query := `
+    INSERT INTO refresh_tokens VALUES ($1, $2, $3, FALSE)
+    ON CONFLICT (email) DO UPDATE SET (hashed_token, valid_until, revoked) = (excluded.hashed_token, excluded.valid_until, excluded.revoked)
+    `
+	_, err = s.database.Exec(query, email, hashed_token, expiry.Time)
+	if err != nil {
+		return fmt.Errorf("inserting refresh token: %w", err)
+	}
+	return nil
+}
+
+func (s *Server) UpsertPasswordActionToken(email, actionType string, hashedToken []byte, validUntil time.Time) error {
+	query := `
+    INSERT INTO password_action_tokens (email, action_type, hashed_token, valid_until, used)
+    VALUES ($1, $2, $3, $4, FALSE)
+    ON CONFLICT (email, action_type)
+    DO UPDATE SET
+        hashed_token = excluded.hashed_token,
+        valid_until = excluded.valid_until,
+        used = FALSE,
+        used_at = NULL
+    `
+	_, err := s.database.Exec(query, email, actionType, hashedToken, validUntil)
+	if err != nil {
+		return fmt.Errorf("upserting password action token: %w", err)
+	}
+	return nil
+}
+
+func (s *Server) ConsumePasswordActionToken(tx *sql.Tx, hashedToken []byte) (string, string, error) {
+	var email string
+	var actionType string
+	err := tx.QueryRow(`
+        SELECT email, action_type
+        FROM password_action_tokens
+        WHERE hashed_token = $1 AND used = FALSE AND valid_until > NOW()
+        FOR UPDATE
+    `, hashedToken).Scan(&email, &actionType)
+	if err == sql.ErrNoRows {
+		return "", "", ErrInvalidPasswordActionToken
+	}
+	if err != nil {
+		return "", "", fmt.Errorf("querying password action token: %w", err)
+	}
+
+	_, err = tx.Exec(`
+        UPDATE password_action_tokens
+        SET used = TRUE, used_at = NOW()
+        WHERE email = $1 AND action_type = $2
+    `, email, actionType)
+	if err != nil {
+		return "", "", fmt.Errorf("marking password action token used: %w", err)
+	}
+	return email, actionType, nil
+}
+
+func (s *Server) UpdatePasswordByEmail(tx *sql.Tx, email string, hashedPassword []byte) error {
+	employeeRes, err := tx.Exec(`
+        UPDATE employees
+        SET password = $1, updated_at = NOW()
+        WHERE email = $2
+    `, hashedPassword, email)
+	if err != nil {
+		return fmt.Errorf("updating employee password: %w", err)
+	}
+	employeeRows, err := employeeRes.RowsAffected()
+	if err != nil {
+		return fmt.Errorf("reading employee affected rows: %w", err)
+	}
+	if employeeRows > 0 {
+		return nil
+	}
+
+	clientRes, err := tx.Exec(`
+        UPDATE clients
+        SET password = $1, updated_at = NOW()
+        WHERE email = $2
+    `, hashedPassword, email)
+	if err != nil {
+		return fmt.Errorf("updating client password: %w", err)
+	}
+	clientRows, err := clientRes.RowsAffected()
+	if err != nil {
+		return fmt.Errorf("reading client affected rows: %w", err)
+	}
+	if clientRows == 0 {
+		return fmt.Errorf("user not found for email")
+	}
+	return nil
+}
+
+func (s *Server) RevokeRefreshTokensByEmail(tx *sql.Tx, email string) error {
+	_, err := tx.Exec(`UPDATE refresh_tokens SET revoked = TRUE WHERE email = $1`, email)
+	if err != nil {
+		return fmt.Errorf("revoking refresh tokens: %w", err)
+	}
+	return nil
+}
