+++ internal/user/employee_test.go
@@ -1,0 +1,3 @@
+package user
+
+// TODO: unit testovi za employee.go
